+++ komposer/core/extra_manifest.py
@@ -1,10 +1,13 @@
 import itertools
-from typing import Literal
+from collections.abc import Mapping, Sequence
+from typing import Any, Union
 
-from pydantic import parse_obj_as
-
+from komposer.exceptions import (
+    ExtraManifestInvalidYamlError,
+    ExtraManifestMissingMetadataError,
+    ExtraManifestMissingNameError,
+)
 from komposer.types import kubernetes
-from komposer.types.base import CamelCaseImmutableBaseModel
 from komposer.types.cli import Context
 from komposer.utils import load_yaml
 
@@ -16,21 +19,15 @@ def _komposer_service_prefix_value(context: Context) -> str:
 komposer_env_variables_map = {"${KOMPOSER_SERVICE_PREFIX}": _komposer_service_prefix_value}
 
 
-class ExtraManifest(CamelCaseImmutableBaseModel):
-    api_version: Literal["v1"] = "v1"
-    kind: Literal["List"] = "List"
-    items: list[kubernetes.Item] = []
-
-
-def update_item_metadata_labels(item: dict, labels: dict) -> None:
+def update_item_metadata_labels(item: Mapping, labels: Mapping) -> None:
     item["metadata"].setdefault("labels", {}).update(labels)
 
 
-def update_item_metadata_name(item: dict, manifest_prefix: str) -> None:
+def update_item_metadata_name(item: Mapping, manifest_prefix: str) -> None:
     item["metadata"]["name"] = f"{manifest_prefix}-{item['metadata']['name']}"
 
 
-def update_item_env_configmapkeyref_name(item: dict, manifest_prefix: str) -> None:
+def update_item_env_configmapkeyref_name(item: Mapping, manifest_prefix: str) -> None:
     containers = item.get("spec", {}).get("template", {}).get("spec", {}).get("containers", [])
     envs = itertools.chain(*(container.get("env", []) for container in containers))
     config_map_key_refs = (env.get("valueFrom", {}).get("configMapKeyRef", {}) for env in envs)
@@ -47,6 +44,42 @@ def replace_env_variables(context: Context, extra_manifest_str: str) -> str:
     return extra_manifest_str
 
 
+def get_items_from_extra_manifest(extra_manifest_raw: Union[Mapping, Sequence]) -> list[dict]:
+    # Returns empty list if null
+    if extra_manifest_raw is None:
+        return []
+
+    # Returns if it's a list
+    if isinstance(extra_manifest_raw, list):
+        return extra_manifest_raw
+
+    # Returns if it's a object
+    if isinstance(extra_manifest_raw, dict):
+        if extra_manifest_raw.get("kind") == "List":
+            return extra_manifest_raw.get("items", [])
+
+        return [extra_manifest_raw]
+
+    # Not uspported
+    raise ExtraManifestInvalidYamlError(
+        f"Unsupported extra manifest type: {type(extra_manifest_raw)}"
+    )
+
+
+def ensure_metadata_is_present(extra_manifest_items: Sequence[Mapping[str, Any]]) -> None:
+    for extra_manifest_item in extra_manifest_items:
+        metadata = extra_manifest_item.get("metadata")
+        if metadata is None:
+            raise ExtraManifestMissingMetadataError(
+                f"Missing metadata in extra manifest item: {extra_manifest_item}"
+            )
+
+        if "name" not in metadata:
+            raise ExtraManifestMissingNameError(
+                f"Missing metadata.name in extra manifest item: {extra_manifest_item}"
+            )
+
+
 def load_extra_manifest(context: Context) -> list[dict]:
     # Skip if no extra manifest
     if context.extra_manifest_path is None:
@@ -61,14 +94,12 @@ def load_extra_manifest(context: Context) -> list[dict]:
     # Parse manifest
     extra_manifest_raw = load_yaml(extra_manifest_str)
 
-    # just validate that the format of the file
-    parse_obj_as(ExtraManifest, extra_manifest_raw)
+    # Just validate that the format of the file
+    extra_manifest_items = get_items_from_extra_manifest(extra_manifest_raw)
 
-    # this is necessary because we need to manipulate the original item
-    # and the item can be undefined in our list of types
-    extra_manifest_items: list[dict] = extra_manifest_raw.get("items", [])
+    ensure_metadata_is_present(extra_manifest_items)
 
-    # Update metadate labels and names and env's configmap names
+    # Update metadata labels and names and env's configmap names
     labels = kubernetes.Metadata.labels_from_context(context)
     manifest_prefix = context.manifest_prefix