+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # v0.1.1
 
 - made the `--project-name` parameter optional
+- relaxed the format of the Kubernetes file passed to the `--extra-minfest` option
 
 # v0.1.0
 
 - Initial version
-= supports only the `GET /loki/api/v1/query` endpoint